requirements.txt
pymongo==4.1.1
Flask==2.1.1
requests==2.27.1
IMDbPY==2021.4.18
Jinja2==3.1.1

